refactor: separate integration tests from unit test suite
Move integration tests to proper location and clean up test organization:
- Move 5 Unix socket/threading integration tests from src/server_tests.rs
  to tests/listener_wake_pattern.rs and tests/server_shutdown.rs
- Remove mock-based tests from unit test suite
- Remove server_tests.rs module reference from lib.rs
- Clean up client integration tests to keep only legitimate unit tests

Result: 115 fast unit tests (isolated, no external dependencies) and 5+ real
integration tests (actual Unix sockets, threading, system integration).
Eliminates slow/flaky tests from unit test suite.
