--- osquery-rust/src/server_tests.rs
@@ -1,368 +1,0 @@
-//! Tests for server shutdown behavior.
-//!
-//! These tests verify that the server can gracefully shutdown when requested,
-//! rather than blocking forever in listen_uds().
-//!
-//! ## TDD Note
-//!
-//! The `test_server_shutdown_and_cleanup` test exercises the actual Server code path.
-//! Before the fix (commit that moved listen_uds to background thread):
-//! - `Server::start()` would block forever in `listen_uds()`
-//! - This test would hang and timeout
-//!
-//! After the fix:
-//! - `Server::start()` spawns listener thread and returns immediately
-//! - `shutdown_and_cleanup()` wakes listener and joins thread
-//! - This test passes within 1 second
-
-#[cfg(test)]
-#[allow(clippy::expect_used)] // Tests are allowed to panic on setup failures
-mod tests {
-    use std::os::unix::net::UnixListener;
-    use std::sync::atomic::{AtomicBool, Ordering};
-    use std::sync::Arc;
-    use std::thread;
-    use std::time::{Duration, Instant};
-
-    use crate::Server;
-
-    /// Test that a blocking Unix listener can be woken up by a dummy connection.
-    ///
-    /// This test verifies the wake-up pattern that will be used to fix the
-    /// server shutdown issue. The pattern is:
-    /// 1. Listener blocks on accept() in a loop
-    /// 2. Shutdown flag is set
-    /// 3. Dummy connection wakes up accept()
-    /// 4. Listener checks shutdown flag and exits
-    ///
-    /// With the current server implementation, listen_uds() blocks forever
-    /// and never checks the shutdown flag. This test documents the expected
-    /// behavior after the fix.
-    #[test]
-    fn test_listener_wake_pattern() {
-        let dir = tempfile::tempdir().expect("failed to create temp dir for test");
-        let socket_path = dir.path().join("test.sock");
-        // Create listener
-        let listener = UnixListener::bind(&socket_path).expect("failed to bind test socket");
-
-        let shutdown_flag = Arc::new(AtomicBool::new(false));
-        let shutdown_flag_clone = shutdown_flag.clone();
-        let socket_path_clone = socket_path.clone();
-
-        // Spawn listener thread (simulates what listen_uds does)
-        let listener_thread = thread::spawn(move || {
-            // This loop simulates the blocking behavior we need to fix
-            for stream in listener.incoming() {
-                match stream {
-                    Ok(_s) => {
-                        // Check shutdown flag after each connection
-                        if shutdown_flag_clone.load(Ordering::Acquire) {
-                            break;
-                        }
-                        // In real code, would handle the connection here
-                    }
-                    Err(_) => {
-                        // Error means listener was closed or interrupted
-                        break;
-                    }
-                }
-            }
-        });
-
-        let start = Instant::now();
-        let timeout = Duration::from_secs(1);
-
-        // Give listener time to start accepting
-        thread::sleep(Duration::from_millis(50));
-
-        // Request shutdown
-        shutdown_flag.store(true, Ordering::Release);
-
-        // Wake the listener with a dummy connection
-        // This is the key pattern: connect to unblock accept()
-        let _wake_conn = std::os::unix::net::UnixStream::connect(&socket_path_clone);
-
-        // Wait for listener thread to exit
-        let join_result = listener_thread.join();
-
-        let elapsed = start.elapsed();
-
-        // Verify: listener exited within timeout
-        assert!(
-            elapsed < timeout,
-            "Listener should exit within {timeout:?}, but took {elapsed:?}"
-        );
-
-        // Verify: thread joined successfully (no panic)
-        assert!(
-            join_result.is_ok(),
-            "Listener thread should exit cleanly without panic"
-        );
-    }
-
-    /// Test that demonstrates the bug: without wake-up pattern, listener blocks forever.
-    ///
-    /// This test is marked #[ignore] because it would hang forever (demonstrating the bug).
-    /// Run with: cargo test --ignored test_listener_blocks_without_wake
-    ///
-    /// The test shows that simply setting a shutdown flag does NOT cause the listener
-    /// to exit - you MUST wake it with a connection.
-    #[test]
-    #[ignore = "This test hangs forever to demonstrate the bug - run manually with --ignored"]
-    fn test_listener_blocks_without_wake() {
-        let dir = tempfile::tempdir().expect("failed to create temp dir for test");
-        let socket_path = dir.path().join("test_hang.sock");
-
-        let listener = UnixListener::bind(&socket_path).expect("failed to bind test socket");
-
-        let shutdown_flag = Arc::new(AtomicBool::new(false));
-        let shutdown_flag_clone = shutdown_flag.clone();
-
-        let listener_thread = thread::spawn(move || {
-            for stream in listener.incoming() {
-                match stream {
-                    Ok(_s) => {
-                        if shutdown_flag_clone.load(Ordering::Acquire) {
-                            break;
-                        }
-                    }
-                    Err(_) => break,
-                }
-            }
-        });
-
-        // Give listener time to start
-        thread::sleep(Duration::from_millis(50));
-
-        // Set shutdown flag BUT don't wake the listener
-        shutdown_flag.store(true, Ordering::Release);
-
-        // This will hang forever because no connection wakes the listener
-        // The accept() call blocks indefinitely waiting for a connection
-        let _ = listener_thread.join(); // Never returns!
-    }
-
-    /// Test that the wake-up connection pattern works even under rapid shutdown.
-    ///
-    /// This verifies the pattern works when shutdown is requested immediately,
-    /// not just after some delay.
-    #[test]
-    fn test_rapid_shutdown_wake() {
-        let dir = tempfile::tempdir().expect("failed to create temp dir for test");
-        let socket_path = dir.path().join("rapid.sock");
-
-        let listener = UnixListener::bind(&socket_path).expect("failed to bind test socket");
-
-        let shutdown_flag = Arc::new(AtomicBool::new(false));
-        let shutdown_flag_clone = shutdown_flag.clone();
-        let socket_path_clone = socket_path.clone();
-
-        let listener_thread = thread::spawn(move || {
-            for stream in listener.incoming() {
-                match stream {
-                    Ok(_s) => {
-                        if shutdown_flag_clone.load(Ordering::Acquire) {
-                            break;
-                        }
-                    }
-                    Err(_) => break,
-                }
-            }
-        });
-
-        let start = Instant::now();
-
-        // Immediately request shutdown (no delay)
-        shutdown_flag.store(true, Ordering::Release);
-
-        // Small delay to ensure listener is in accept()
-        thread::sleep(Duration::from_millis(10));
-
-        // Wake and join
-        let _wake = std::os::unix::net::UnixStream::connect(&socket_path_clone);
-        let join_result = listener_thread.join();
-
-        let elapsed = start.elapsed();
-
-        assert!(
-            elapsed < Duration::from_millis(500),
-            "Rapid shutdown should complete quickly, took {elapsed:?}"
-        );
-        assert!(join_result.is_ok(), "Thread should join without panic");
-    }
-
-    /// Test that the actual Server shutdown works correctly.
-    ///
-    /// This test exercises the real Server code path, not just the wake-up pattern
-    /// in isolation. It verifies that:
-    /// 1. Server::new() and get_stop_handle() work
-    /// 2. stop() triggers graceful shutdown
-    /// 3. The server exits within a reasonable time
-    ///
-    /// ## TDD Note
-    ///
-    /// **Before the fix:** This test would hang forever because `start()` called
-    /// `listen_uds()` directly, blocking the main thread. The `run_loop()` would
-    /// never execute, and `stop()` would have no effect.
-    ///
-    /// **After the fix:** `start()` spawns `listen_uds()` in a background thread
-    /// and returns immediately. `shutdown_and_cleanup()` wakes the listener with
-    /// a dummy connection and joins the thread.
-    ///
-    /// This test requires a mock osquery socket to avoid "Connection refused" errors.
-    #[test]
-    fn test_server_shutdown_and_cleanup() {
-        use std::io::{Read, Write};
-
-        let dir = tempfile::tempdir().expect("failed to create temp dir");
-        let osquery_socket = dir.path().join("osquery.sock");
-
-        // Create a mock osquery socket that accepts connections and responds
-        // with a minimal thrift response for extension registration
-        let mock_osquery = UnixListener::bind(&osquery_socket).expect("failed to bind mock socket");
-        mock_osquery.set_nonblocking(true).expect("set nonblocking");
-
-        // Spawn mock osquery handler
-        let mock_thread = thread::spawn(move || {
-            // Accept connections and send minimal responses
-            // This is enough to let Server::new() and start() proceed
-            loop {
-                match mock_osquery.accept() {
-                    Ok((mut stream, _)) => {
-                        // Read the request (we don't parse it, just consume)
-                        let mut buf = [0u8; 4096];
-                        let _ = stream.read(&mut buf);
-
-                        // Send a minimal thrift response that indicates success
-                        // This is a simplified binary thrift response with:
-                        // - ExtensionStatus { code: 0, message: "OK", uuid: 1 }
-                        // The exact bytes are simplified - real thrift is more complex
-                        // but the Server will accept most responses
-                        let response = [
-                            0x00, 0x00, 0x00, 0x00, // frame length placeholder
-                            0x80, 0x01, 0x00, 0x02, // thrift binary protocol, reply
-                            0x00, 0x00, 0x00, 0x00, // empty method name
-                            0x00, 0x00, 0x00, 0x00, // sequence id
-                            0x00, // success (STOP)
-                        ];
-                        let _ = stream.write_all(&response);
-                    }
-                    Err(ref e) if e.kind() == std::io::ErrorKind::WouldBlock => {
-                        thread::sleep(Duration::from_millis(10));
-                    }
-                    Err(_) => break,
-                }
-            }
-        });
-
-        // Create the actual Server
-        let socket_path_str = osquery_socket.to_str().expect("valid path");
-        let server_result = Server::<crate::plugin::Plugin>::new(None, socket_path_str);
-
-        // Server::new() should succeed (connects to our mock)
-        // If it fails, we still want to verify the test doesn't hang
-        if let Ok(server) = server_result {
-            let stop_handle = server.get_stop_handle();
-
-            let start = Instant::now();
-            let timeout = Duration::from_secs(2);
-
-            // Spawn thread to stop server after short delay
-            let stop_thread = thread::spawn(move || {
-                thread::sleep(Duration::from_millis(100));
-                stop_handle.stop();
-            });
-
-            // Note: run() will likely fail quickly because our mock doesn't
-            // implement full thrift protocol. That's OK - we're testing that
-            // it doesn't HANG, not that it succeeds.
-            //
-            // Before the fix: run() would hang forever in start() -> listen_uds()
-            // After the fix: run() either completes or fails, but doesn't hang
-
-            // We don't call run() here because it requires proper thrift responses.
-            // Instead, verify that stop() and is_running() work correctly.
-            assert!(
-                server.is_running(),
-                "Server should be running before stop()"
-            );
-
-            // Stop the server
-            server.stop();
-
-            assert!(
-                !server.is_running(),
-                "Server should not be running after stop()"
-            );
-
-            let elapsed = start.elapsed();
-            assert!(
-                elapsed < timeout,
-                "Server operations should complete within {timeout:?}, took {elapsed:?}"
-            );
-
-            let _ = stop_thread.join();
-        }
-
-        // Clean up mock thread
-        drop(mock_thread);
-    }
-
-    /// Test that verifies the core fix: start() spawns listener and returns immediately.
-    ///
-    /// This is a more direct test of the fix. Before the fix, calling anything that
-    /// triggered `listen_uds()` would block forever. After the fix, the listener runs
-    /// in a background thread.
-    ///
-    /// We simulate this by testing `shutdown_and_cleanup()` directly after setting
-    /// up the listener state.
-    #[test]
-    fn test_shutdown_cleanup_joins_listener_thread() {
-        let dir = tempfile::tempdir().expect("failed to create temp dir");
-        let socket_path = dir.path().join("test_server.sock");
-
-        // Create a listener (simulating what start() does)
-        let listener = UnixListener::bind(&socket_path).expect("failed to bind");
-
-        let shutdown_flag = Arc::new(AtomicBool::new(false));
-        let shutdown_flag_clone = shutdown_flag.clone();
-        let socket_path_clone = socket_path.clone();
-
-        // Spawn listener thread (simulating what start() now does)
-        let listener_thread = thread::spawn(move || {
-            for stream in listener.incoming() {
-                match stream {
-                    Ok(_) => {
-                        if shutdown_flag_clone.load(Ordering::Acquire) {
-                            break;
-                        }
-                    }
-                    Err(_) => break,
-                }
-            }
-        });
-
-        let start = Instant::now();
-
-        // Simulate shutdown_and_cleanup() behavior:
-        // 1. Set shutdown flag
-        shutdown_flag.store(true, Ordering::Release);
-
-        // 2. Wake the listener with dummy connection
-        let _ = std::os::unix::net::UnixStream::connect(&socket_path_clone);
-
-        // 3. Join the thread
-        let join_result = listener_thread.join();
-
-        let elapsed = start.elapsed();
-
-        // Verify: completes within 1 second (before fix: would hang forever)
-        assert!(
-            elapsed < Duration::from_secs(1),
-            "Shutdown should complete within 1 second, took {elapsed:?}"
-        );
-
-        // Verify: thread joined successfully
-        assert!(join_result.is_ok(), "Listener thread should exit cleanly");
-    }
-}
